This tool retrieves a list of bug numbers from the output of 'git log'
It then grabs various descriptive columns from bugzilla and formats and
inserts them into a release note template. Next it retrieves a list of
commit authors from 'git log,' sorts them alphabetically by last name
(or tries to), and inserts them into the template.

At this point the template is not really a template nor is this tool
paramatized enought to avoid code changes specific to the present
release.

It is very hackish... you've been warned.