A tool to lookup commits which need to be cherry-picked
The aim of this tool is to parse the output of 'git cherry,' look up and
determine which bugs are enh requests and which are not, automagically
cherry-pick those which are not, and dump a list of enh requests for
manual review.

It also creates a list of commit log entries which have no associated bug
number for manual review.

WARNING: This script is incomplete with this first commit.