Add full description to enhancements
Displaying only the short description is not enough for enhancements.
This patch retrieve the "comment 0", that is the complete description of what the patch does

This information cannot be retrieved from the initial search.
Thus, we need to query with bugz to retrieve everything about, and extract the comment 0

As it's a slow step and the comment 0 can't be updated, the result is stored in a Storable
Thus, if the query is relaunched, it won't query bugzilla again where unneeded.

To use this, pass your bugzilla login & password in parameters